refactor of all the paintings

+++ src/main/java/com/museum/gestionale/Dao/PaintingDao.java
@@ -8,8 +8,8 @@ import javax.management.Query;
 import java.awt.print.Pageable;
 import java.util.List;
 
-public interface PaintingDao extends CrudRepository<Painting, Integer> {
-	Painting findById(int id);
+public interface PaintingDao extends CrudRepository<Painting, Long> {
+	Painting findById(long id);
 	List<Painting> findAll();
 	List<Painting> findAllByTags(Tag tags);
 	List<Painting> findByName(String name);
